Move Lua unit attacks metatable to own file
